Do not add "grunt-cli" to client/package.json.
It is not really a requirement, but it is needed in order to build the client.
I am solving this issue by:
- Installing globally on a travis machine,
- Installling locally for building the debian package,
- Let the Globaleaks developers do whatevers they want on their system (either
  have it globally or locally, but not forcing on any of the two).
run the local version of grunt, even if that raises a warning.
Maybe there are better ways to do this, but as of now I don't see any 'grunt'
binary shipped by debian, and we really must build glclient without beeing
root..