Introduce packer user variable `deb`.
The introduction of this variable is due to the fact that packer does not report
in the official documentation any way to guess globaleaks version, and find a
globaleaks debian package automagically. Therefore, from now on the user will
have to specify this variable pointing to the globaleaks .deb file generated by
dpkg. E.g.:

$ packer build -var deb=../../../globaleaks_2.60.30_all.deb template.json