Add client side javascript code coverage report for end to end protractor tests.
With these changes it's possible to see the coverage of the client side javascript code from the end to end (protractor) tests. These results are sent to coveralls and added up to the coverage results of the python (trial) unit tests to have a global coverage mesure.

Change CI (travis) to run client end to end tests for coverage together with unit tests on first job and to send coverage info to coveralls iaggregates in one step. For that, it is necessary to generate coveralls json for client coverage separately.