Refactor mail routines in relation to detected infinite jobs happening in case of mail failures
It seems to happen that the twisted deferred chain is not calling some callbacks whenever particular configuration are done.
E.g. this happens whenever the server is using TLS but the user conffigure to use SSL; what happens on this condition is that twisted closes the connection without providing us a callback.

With this commit i've:
- improved the resiliency by simpliyfing the SMTP related code to the min.
- acted a regression on some of the improvals of #214 that where already integrated; after a deep testing all the day long i found that they were not working properly at all.
- implemented a timeout guard on the sending on a single mail, now capped to 2 minutes
- i've improved the security of the kind of cyphers used using SMTP authentication deprecating SSLv3 as done in tor2web (i already did int in past but the solution was incomplete.