Fix get_future_epoch localization.
get_future_epoch used time.timezone to incorrectly compute seconds since
the epoch. The test was only failing for me, because time.timezone
returned +18000 – a 5 hour time difference. These changes fix the unit test as
well. However, this function is used nowhere in the code so it might be
better to delete it anyhow....!