Remove CSSMin standing on the shoulders of giants
This commit is one of a series.

The series started with https://github.com/globaleaks/GlobaLeaks/issues/1417
when we decided to avoid uglifying contents in order to possibly avoid
minification bugs, be more reproducible and for adding more debugging options at runtime.

With this commit the same change is applied to minification of CSSs;
There is no need for duing that; we can rely on upstream minified versions that are tagged
and tested by largers communities.

For what relates our code the codebase is still so small that we have no valid need for
minification of the JSs nor of the CSSs files

In addition the commit remove the grunt-concat-css library that we were not using at all.