Remove logging stmts and add reload after SetLang()
Some comment and spacing edits as well. @evilaliv3, this whole pull
request is ready to merge.

There is one bug that occurs when default lang is set to something other
than english and the client can sniff a supported language. In this
situation the default strings in `node` are bound once to the wrong
language (english in all cases).