Added file and answers signing and a glbcWhistleblower factory
A careful looks needs to be taken at the structure of the keys used to
sign the pgp messages. Currently, it looks as if openpgp does NOT enforce
rsa_sign_only:
https://github.com/openpgpjs/openpgpjs/blob/master/src/crypto/signature.js#L30-L35

glbcWhistleblower handles all of the encryption/decryption and
signing and verification that the WB takes in the app.

glbcCipherLib encryptMsg/decryptMsg will be removed.