Flag switched from forced_pass_change to ccrypto_key_public
This commit deprecates the use of forced_password_change as the
flag for initialization of client side keys. Instead the existence of
the public key itself is used to indicate to the client that it must be
configured.

Additional logic catches the cases when a password change before a
user's first login has been 'unforced' now ignores the flag.