Implement migration for #1739
In the migration to db v_34 we add an archive of appdata_v2_62_8.json to
use as the default translations coming from the 'old_db'. The real
defaults were placed in the db when the application was first started.

The preference for the platform will be to use whatever value is in the
db. By setting the default value for the configuration to the current app_data
it leaves the possibility to reset the values to the 'defaults' at later
time.

With all this said, the schema is set to change as we intergrate this
into the larger config refactor.