Apply revision of implementation for #2017
This patch corrects the tracking of the qury limiting the tracking to the time between the database lock is taken and the actual release.
The fix is necessary in order to limit the count of the execution time per query, without counting the time the query waited in the queue.