Add LetsEncrypt HTTPS enrollment UI into wizard panels
The idea to support the behavior for #2019 is simple. After the wizard
has generated a session, use the normal HTTPS CfgCntrl in the wizard
interface to do what it normally does. There are only a 3 pieces of
configuration that a parent controller most define via $scope.https_.

This commit is not definitive support b/c there are surely issues to
handle.