Fix implementation of x509_cert_check_sched in relation to HTTPS processes restart
The main fix implemented is the change of the paradigm of the job from threaded to no threaded;
this way the job could correctly make use of the reactor yields and inlineCallbacks as required.

The commit does also implement significative simplifications arounds queries reducing their number from 3 to 1 that
due to the fact that the job is not threaded become fundamental requirement.