Handle typical special situation of demo setup and recipients turnover
During initial setup and on recipients' turnover it is likely that recipients may have not perfomed
initial login and that encryptions keys have still not being generated.
In this special condition, in the case that no recipient is in possession of encryption keys it is a
cceptable to deliver the submission not encrypted instead of loosing it.