Address usability issue #3098
To handle this peculiar usability issue we decided that is acceptable for a user to set an expiration date exceeding the default data retention policy by one additional period.

E.g. if the policy is set to one year the user when postponing will be able to choose a date two years ahead.