Implement client session idle check and keep alive requests
This commit introduces client session idle check and keep alive requests.

By using the library ng-idle the software will be able to detect user presence and automatically renew the user session.

This will guarantee tha if the user is using the application by scrolling the page or typing answers to the submission questionnaire the session will be kept open.

The ticket addresses both issue #603 and #2842