Documentation/Docker: Fix Debian release detection
In install.sh, a comment (Buster) and a fixed output string (Bullseye)
were inconsistent to the actual default (bookworm). Plus, in a Docker
build using Debian 12-slim, the lsb_release command is not present,
resulting in DISTRO_CODENAME to be 'unknown'. Fixed by installing the
package lsb-release in Dockerfile before the install script is
executed.

Signed-off-by: Olav Seyfarth <olav@seyfarth.de>