Deprecate file preview based on data uri
The file previww implementation could not work since the Content Security Policy has been improved deprecating the usage of data: uri and will need to be reimplemented on the base of the secure file viewer.

The benefits of such an implementation would consists in the possibility to implement preview for PDF and other types of files other than only images and to possibly show metadata to users / enable them to redact part of the file before of the upload.